clean frontend folder
